feat(config): Handle RUN_CMD that contains spaces.
When the run command is one that has subcommands, it is sometimes
easier to set the set of subcommands in the base run environment.

However when using child_process.spawn, this does not work unless
those same subcommands are passed in as args. This will split on
the whitespace and pass the subcommands as the initial args.